+++ example/lib/tile/extension_warning_tile_stub.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
 
 class ExtensionWarningTile extends StatelessWidget {
-  const ExtensionWarningTile({super.key, this.hideWhileLoadingAndReady = false});
+  const ExtensionWarningTile({
+    super.key,
+    this.hideWhileLoadingAndReady = false,
+  });
 
   final bool hideWhileLoadingAndReady;